02_invoice_reader_prototype/requirements.txt
# Core
pandas>=2.0
numpy>=1.24

# Database
SQLAlchemy>=2.0

# Dates
python-dateutil>=2.8

# PDF + images
pdf2image>=1.16
Pillow>=10.0

# Gemini API
google-generativeai>=0.5

# Environment variables
python-dotenv>=1.0
